Update : Run Test / Patch  (#issues29)
